% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic_suitability.R
\name{abiotic_suitability}
\alias{abiotic_suitability}
\title{Abiotic suitability risk mapping layer}
\usage{
abiotic_suitability(
  sdm_output,
  template,
  normalize = FALSE,
  filename = "",
  ...
)
}
\arguments{
\item{sdm_output}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the SDM output spatial layer.}

\item{template}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the risk mapping template spatial layer.}

\item{normalize}{Logical indicating if the combined cells should be
normalized, i.e. set to a value 0-1 based on cell-wise minimum and
maximum values, i.e. \code{(value - min)/(max - min)}. Default =
\code{FALSE}.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the conformed abiotic
suitability layer.
}
\description{
Builds an abiotic suitability spatial layer by conforming an existing
(global) species distribution model (SDM) output layer to the spatial
configuration (CRS, extent, and resolution) of the selected risk mapping
template layer.
}
\note{
Informed by various functions in
\code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
