% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_categories.R
\name{aggregate_categories}
\alias{aggregate_categories}
\alias{aggregate_categories.Raster}
\alias{aggregate_categories.SpatRaster}
\title{Aggregate categories within a spatial layer}
\usage{
aggregate_categories(
  x,
  y,
  categories = NULL,
  selected = NULL,
  binarize = TRUE,
  ...
)

\method{aggregate_categories}{Raster}(
  x,
  y,
  categories = NULL,
  selected = NULL,
  binarize = TRUE,
  ...
)

\method{aggregate_categories}{SpatRaster}(
  x,
  y,
  categories = NULL,
  selected = NULL,
  binarize = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer of category values to be aggregated.}

\item{y}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer with the spatial resolution (and
configuration) to aggregate (conform) to.}

\item{categories}{A numeric vector of all possible category values. If this
is left \code{NULL} (default) the category values are extracted from the
\code{x} layer via possible pre-existing categories (if present) or via
calculating unique cell values (slower).}

\item{selected}{A numeric vector of all possible category values. If this
is left \code{NULL} (default) no categories will be selected.}

\item{binarize}{Logical indicating if the aggregated cells should be
binarized, i.e. set to 1 for cells containing selected category values.
Default = TRUE.}

\item{...}{Additional parameters (unused).}
}
\value{
A \code{raster::RasterLayer} or \code{terra::SpatRaster} object
(as per \code{pathway_layers}) containing the conformed layer.
}
\description{
Aggregate the categories within a spatial layer to a courser resolution,
that of another (template) layer, based on the presence of each of the
user-selected categories in each cell. The resulting aggregated layer cells
may be binarized to one or zero, or indicate the proportion of selected
categories present in the cell.
}
\note{
Informed by various functions in
\code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
