% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_layer.R
\name{aggregate_layer}
\alias{aggregate_layer}
\alias{aggregate_layer.Raster}
\alias{aggregate_layer.SpatRaster}
\title{Aggregate spatial layer}
\usage{
aggregate_layer(
  x,
  y,
  use_fun = c("mean", "max", "min", "median", "sum", "modal", "union"),
  use_method = c("auto", "bilinear", "near", "sum"),
  platform = FALSE,
  filename = "",
  ...
)

\method{aggregate_layer}{Raster}(
  x,
  y,
  use_fun = c("mean", "max", "min", "median", "sum", "modal", "union"),
  use_method = c("auto", "bilinear", "near", "sum"),
  platform = FALSE,
  filename = "",
  ...
)

\method{aggregate_layer}{SpatRaster}(
  x,
  y,
  use_fun = c("mean", "max", "min", "median", "sum", "modal", "union"),
  use_method = c("auto", "bilinear", "near", "sum"),
  platform = FALSE,
  filename = "",
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer to be aggregated.}

\item{y}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer with the spatial resolution (and
configuration) to aggregate (conform) to.}

\item{use_fun}{One of \code{"mean"}, \code{"max"}, \code{"min"},
\code{"median"}, \code{"sum"}, \code{"modal"}, or \code{"union"}:
\code{1 - prod(1 - x)}.}

\item{use_method}{Use method when projecting and/or re-sampling. One of
\code{"auto"} (uses "near" for categorical rasters else "bilinear"),
\code{"bilinear"}, \code{"near"}, or \code{"sum"}.
Default = \code{"auto"}.}

\item{platform}{Logical indicating function is to be run in a platform
environment requiring workaround code. Default = \code{FALSE}.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the aggregated layer.
}
\description{
Aggregate the spatial resolution of a fine grain layer to that of another
(template) layer having a course resolution. Other spatial configuration
(CRS and extent) are checked and adjusted to ensure they also match that of
the template.
}
\note{
Informed by various functions in
\code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
