% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival_likelihood.R
\name{arrival_likelihood}
\alias{arrival_likelihood}
\title{Pest arrival likelihood risk mapping layer}
\usage{
arrival_likelihood(pathway_likelihood_layers, filename = "", ...)
}
\arguments{
\item{pathway_likelihood_layers}{A multi-layer \code{raster::Raster*} or
\code{terra::SpatRaster} object representing the pathway likelihood layers
to combine (union) to form a pest arrival likelihood layer.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the combined pest
arrival likelihood layer.
}
\description{
Builds a arrival likelihood spatial layer by combining pest pathway
likelihood layers (via layer union).
}
\note{
Generalized modified version of corresponding layer-combining
functions in \code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
