% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotic_suitability.R
\name{biotic_suitability}
\alias{biotic_suitability}
\title{Biotic suitability risk mapping layer}
\usage{
biotic_suitability(biotic_layers, filename = "", ...)
}
\arguments{
\item{biotic_layers}{A multi-layer \code{raster::Raster*} or
\code{terra::SpatRaster} object representing the biotic (host) spatial
layers to combine (multiply together) to form a biotic suitability layer.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the combined biotic
suitability layer.
}
\description{
Builds a biotic suitability spatial layer by combining (multiplying) biotic
(host) layers.
}
\note{
Generalized modified version of corresponding layer-combining
functions in \code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
