% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffered_hull_layer.R
\name{buffered_hull_layer}
\alias{buffered_hull_layer}
\alias{buffered_hull_layer.Raster}
\alias{buffered_hull_layer.SpatRaster}
\title{Calculate a buffered hull layer}
\usage{
buffered_hull_layer(
  x,
  y,
  hull = c("alpha", "convex", "none"),
  alpha = NULL,
  buffer = NULL,
  filename = "",
  ...
)

\method{buffered_hull_layer}{Raster}(
  x,
  y,
  hull = c("alpha", "convex", "none"),
  alpha = NULL,
  buffer = NULL,
  filename = "",
  ...
)

\method{buffered_hull_layer}{SpatRaster}(
  x,
  y,
  hull = c("alpha", "convex", "none"),
  alpha = NULL,
  buffer = NULL,
  filename = "",
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial configuration (template) for calculating
the buffered hull layer from points \code{y}.}

\item{y}{Point (feature) data as a \code{data.frame} (or \code{matrix})
with WGS84 \emph{lon} and \emph{lat} columns.}

\item{hull}{One of \code{"alpha"}, \code{"convex"}, or \code{"none"} for
the type of hull to fit to the points \code{y}.}

\item{alpha}{Numeric alpha parameter for fitting an alpha hull to the
points \code{y}. The value should be in the same units as the coordinates
(or CRS) of the template layer \code{x}. For example, use degrees if the
coordinates are in WGS84 longitudes and latitudes, otherwise coordinates
are typically in meters.}

\item{buffer}{Numeric buffer distance to place around selected hull, or
around points \code{y} when no hull \code{"none"} is selected. The value
should be in meters when template layer \code{x} coordinates (or CRS) are
in WGS84 longitudes and latitudes, otherwise use the same units as the
coordinates (typically also in meters).}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the calculated buffered
hull layer.
}
\description{
Calculates a spatial layer based on fitting a alpha or convex hull and/or a
distance-based buffer to a series of points (features).
}
\note{
Generalized modified version of corresponding alpha hull function
\code{\href{edmaps}{https://github.com/jscamac/edmaps}::rasterize_range}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
