% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_establish_event_pr.R
\name{calc_establish_event_pr}
\alias{calc_establish_event_pr}
\title{Calculate establishment event probabilities}
\usage{
calc_establish_event_pr(
  n_events,
  p_establish,
  confidence = 0.95,
  n_sims = 1e+05,
  ...
)
}
\arguments{
\item{n_events}{Numeric vector containing lower and upper bounds (CI) for
the number of leakage events.}

\item{p_establish}{Numeric vector containing the lower and upper bounds
(CI) for the probability a leakage event could result in an establishment.}

\item{confidence}{Confidence interval (CI). Default = 0.95.}

\item{n_sims}{Integer. Number of samples to be taken from event and
establishment distributions. Default = 100000.}

\item{...}{Additional parameters (unused).}
}
\value{
A \code{data.frame} containing the possible number of incursions
that may occur, and their corresponding probabilities.
}
\description{
Calculate (via stochastic sampling) the probability of one or more
establishment events occurring as as a function of estimated leakage numbers
and the probability a leakage event could result in establishment.
}
\note{
Modified version of
\code{\href{edmaps}{https://github.com/jscamac/edmaps}::calc_EE}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
