% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pathway_establish_pr.R
\name{calc_pathway_establish_pr}
\alias{calc_pathway_establish_pr}
\alias{calc_pathway_establish_pr.Raster}
\alias{calc_pathway_establish_pr.SpatRaster}
\title{Calculate the pathway-specific probability of pest establishment}
\usage{
calc_pathway_establish_pr(rast, ee_pr, filename = "", ...)

\method{calc_pathway_establish_pr}{Raster}(rast, ee_pr, filename = "", ...)

\method{calc_pathway_establish_pr}{SpatRaster}(rast, ee_pr, filename = "", ...)
}
\arguments{
\item{rast}{A \code{raster::RasterLayer} or \code{terra::SpatRaster} object
containing an estimate of the spatial distribution of the probability of
pest establishment.}

\item{ee_pr}{\code{data.frame} containing overall establishment event
probabilities calculated via \code{\link{calc_establish_event_pr}}.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{raster::RasterLayer} or \code{terra::SpatRaster} object (as
per \code{rast}) containing the calculated pathway-specific
spatially-explicit pest establishment probability estimates.
}
\description{
Calculate the pathway-specific probability of pest establishment for each
raster cell.
}
\note{
Modified version of
\code{\href{edmaps}{https://github.com/jscamac/edmaps}::calc_pathway_pr}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
