% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pathway_pr.R
\name{calc_pathway_pr}
\alias{calc_pathway_pr}
\alias{calc_pathway_pr.Raster}
\alias{calc_pathway_pr.SpatRaster}
\title{Calculate the pathway-specific probability of pest arrival}
\usage{
calc_pathway_pr(rast, EE, ...)

\method{calc_pathway_pr}{Raster}(rast, EE, ...)

\method{calc_pathway_pr}{SpatRaster}(rast, EE, ...)
}
\arguments{
\item{rast}{A \code{raster::RasterLayer} or \code{terra::SpatRaster} object
containing an estimate of the spatial distribution of the probability of
pest arrival.}

\item{EE}{\code{data.frame} containing overall establishment event
probabilities calculated via \code{\link{calc_EE}}.}

\item{...}{Additional parameters (unused).}
}
\value{
A \code{raster::RasterLayer} or \code{terra::SpatRaster} object (as
per \code{rast}) containing the calculated pathway-specific
spatially-explicit pest arrival probability estimates.
}
\description{
Calculate the pathway-specific probability of pest arrival for each raster
cell.
}
\note{
Modified version of
\code{\href{edmaps}{https://github.com/jscamac/edmaps}::calc_pathway_pr}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
