% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conform_layer.R
\name{conform_layer}
\alias{conform_layer}
\alias{conform_layer.Raster}
\alias{conform_layer.SpatRaster}
\title{Conform spatial layer}
\usage{
conform_layer(
  x,
  y,
  normalize = FALSE,
  binarize = FALSE,
  na_strategy = c("zero", "nearest", "retain"),
  use_aggr_fun = c("mean", "max", "min", "median", "sum", "modal", "union"),
  use_method = c("auto", "bilinear", "near", "sum"),
  platform = FALSE,
  filename = "",
  ...
)

\method{conform_layer}{Raster}(
  x,
  y,
  normalize = FALSE,
  binarize = FALSE,
  na_strategy = c("zero", "nearest", "retain"),
  use_aggr_fun = c("mean", "max", "min", "median", "sum", "modal", "union"),
  use_method = c("auto", "bilinear", "near", "sum"),
  platform = FALSE,
  filename = "",
  ...
)

\method{conform_layer}{SpatRaster}(
  x,
  y,
  normalize = FALSE,
  binarize = FALSE,
  na_strategy = c("zero", "nearest", "retain"),
  use_aggr_fun = c("mean", "max", "min", "median", "sum", "modal", "union"),
  use_method = c("auto", "bilinear", "near", "sum"),
  platform = FALSE,
  filename = "",
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer to be conformed.}

\item{y}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer to conform to.}

\item{normalize}{Logical indicating if the conformed cells should be
normalized, i.e. set to a value 0-1 based on cell-wise minimum and
maximum values, i.e. \code{(value - min)/(max - min)}. Default =
\code{FALSE}.}

\item{binarize}{Logical indicating if the conformed layer should be
binarized, i.e. set to 1 for values > 0. Default = \code{FALSE}.}

\item{na_strategy}{One of \code{"zero"}, \code{"nearest"}, or
\code{"retain"}, indicating the strategy for how to set locations where
\code{x} is \code{NA} but the conform template \code{y} has non-NA values.
By default they are set to \code{"zero"}. Alternatively, unwanted
\code{NA} values (e.g. due to mismatches in coastlines) may be set to the
\code{"nearest"} non-NA values, or when applicable, simply \code{"retain"}
\code{NA} values.}

\item{use_aggr_fun}{Use aggregation function when applicable. One of
\code{"mean"}, \code{"max"}, \code{"min"}, \code{"median"}, \code{"sum"},
\code{"modal"}, or \code{"union"}: \code{1 - prod(1 - x)}.
Default = \code{"mean"}.}

\item{use_method}{Use method when projecting and/or re-sampling. One of
\code{"auto"} (uses "near" for categorical rasters else "bilinear"),
\code{"bilinear"}, \code{"near"}, or \code{"sum"}.
Default = \code{"auto"}.}

\item{platform}{Logical indicating function is to be run in a platform
environment requiring workaround code. Default = \code{FALSE}.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the conformed layer.
}
\description{
Conforms the spatial configuration (CRS, extent, and resolution) of a
spatial layer to that of another (template) layer.
}
\note{
Informed by various functions in
\code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
