% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_likelihood.R
\name{cumulative_likelihood}
\alias{cumulative_likelihood}
\alias{cumulative_likelihood.Raster}
\alias{cumulative_likelihood.SpatRaster}
\title{Cumulative likelihood}
\usage{
cumulative_likelihood(
  x,
  output = c("likelihood", "percentage"),
  order_by = c("ascending", "descending"),
  filename = "",
  ...
)

\method{cumulative_likelihood}{Raster}(
  x,
  output = c("likelihood", "percentage"),
  order_by = c("ascending", "descending"),
  filename = "",
  ...
)

\method{cumulative_likelihood}{SpatRaster}(
  x,
  output = c("likelihood", "percentage"),
  order_by = c("ascending", "descending"),
  filename = "",
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial likelihood layer.}

\item{output}{One of ordered cumulative \code{"likelihood"} (default) or
\code{"percentage"} of ordered cumulative total in decile-like groupings.}

\item{order_by}{Either \code{"ascending"} (default) or \code{"descending"}
order of likelihood.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the cumulative
likelihood layer. Percentage outputs produce decile-like categorical
layers.
}
\description{
Calculates the ordered cumulative likelihood of a spatial likelihood layer.
The output may be displayed as cumulative ordered likelihood, or as
percentage of ordered cumulative total in decile-like groupings.
}
