% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight_layer.R
\name{distance_weight_layer}
\alias{distance_weight_layer}
\alias{distance_weight_layer.Raster}
\alias{distance_weight_layer.SpatRaster}
\title{Calculate a distance weight layer}
\usage{
distance_weight_layer(
  x,
  y,
  beta = log(0.5)/200,
  weights = NULL,
  filename = "",
  ...
)

\method{distance_weight_layer}{Raster}(
  x,
  y,
  beta = log(0.5)/200,
  weights = NULL,
  filename = "",
  ...
)

\method{distance_weight_layer}{SpatRaster}(
  x,
  y,
  beta = log(0.5)/200,
  weights = NULL,
  filename = "",
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial configuration (template) for calculating
the distance weight layer from points \code{y}, or a for a pre-calculated
distance layer (assumed when \code{y} is missing).}

\item{y}{Point (feature) data as a \code{data.frame} (or \code{matrix})
with WGS84 \emph{lon} and \emph{lat} columns.}

\item{beta}{Numeric parameter passed to the exponential function:
\code{exp(distance/(1000/beta))}, which transforms cell-to-nearest-point
or pre-calculated distances (m) to distance-weighted cell values. To
generate a distribution that ensures a proportion \emph{p} of a pathway
likelihood is distributed within distance of the nearest point (feature)
or pre-calculated distance \emph{d}, specify \code{beta=log(p)/d} (e.g.
to have 50\% of the likelihood distributed within 200 km use default:
\code{log(0.5)/200}).}

\item{weights}{Optional (default is none) numeric vector of weights for
each point (feature) such that the distance-weighted cell values will be
additionally proportionally weighted. Thus the exponential function
(above) becomes: the maximum of \code{exp(distance/(1000/beta))*weight}
for each cell.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing distance-weighted cell
values.
}
\description{
Calculates a distance-weighted spatial layer based on the proximity of each
cell to a series of points (features), or via a pre-calculated distance
layer. Each cell is weighted via a negative exponential function applied to
the distance of the nearest point or pre-calculated distances.
}
\note{
Generalized modified version of corresponding distance-weight
functions in \code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
