% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_features.R
\name{distribute_features}
\alias{distribute_features}
\alias{distribute_features.Raster}
\alias{distribute_features.SpatRaster}
\title{Distribute features across a raster layer}
\usage{
distribute_features(x, y, vars = NULL, filename = "", ...)

\method{distribute_features}{Raster}(x, y, vars = NULL, filename = "", ...)

\method{distribute_features}{SpatRaster}(x, y, vars = NULL, filename = "", ...)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial configuration (template or mask) for
distributing the feature (polygon) values from \code{y}.}

\item{y}{Feature polygon data as a \code{terra::SpatVector} or a simple
feature \code{sf::sf}. Note that only the feature polygons of \code{y}
contained entirely within the spatial extent of \code{x} will be
distributed.}

\item{vars}{A vector of character variable names corresponding to the
features (columns) in \code{y} to distribute. Only numeric variables are
supported. If \code{NULL}, all numeric variables in \code{y} will be
distributed.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A multi-layer \code{terra::SpatRaster} object containing a spatial
layer for each feature variable from \code{y} specified in \code{vars}.
}
\description{
Distributes feature polygon values across a template or mask raster spatial
layer. The total value assigned to each polygon is divided across all
corresponding raster cells when a template (single value plus NAs) raster
is provided, or across non-zero cells when a mask (zero and non-zero values
plus NAs) raster is provided, which is used as a binary mask (regardless of
the variety of non-zero values).
}
\note{
Generalized modified version of
\code{\href{edmaps}{https://github.com/jscamac/edmaps}} functionality for
distributing polygons across raster cells.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
