% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent_crs.R
\name{equivalent_crs}
\alias{equivalent_crs}
\alias{equivalent_crs.Raster}
\alias{equivalent_crs.SpatRaster}
\title{Checks if CRS of spatial layers is equivalent}
\usage{
equivalent_crs(x, y, ...)

\method{equivalent_crs}{Raster}(x, y, ...)

\method{equivalent_crs}{SpatRaster}(x, y, ...)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing a spatial layer to be compared.}

\item{y}{Another \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing a spatial layer for comparison.}

\item{...}{Additional parameters (unused).}
}
\value{
A logical value indicating if the CRS of spatial layers \code{x} and
\code{y} are equivalent (in some cases).
}
\description{
Checks if the coordinate reference system (CRS) of two spatial layers are
equivalent via removing some redundant PROJ string information. Not
guaranteed to work in every case. A work in progress.
}
