% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establishment_likelihood.R
\name{establishment_likelihood}
\alias{establishment_likelihood}
\title{Pest establishment likelihood risk mapping layer}
\usage{
establishment_likelihood(
  pest_suitability_layer,
  arrival_likelihood_layer,
  filename = "",
  ...
)
}
\arguments{
\item{pest_suitability_layer}{A \code{raster::RasterLayer} or
\code{terra::SpatRaster} object representing the pest suitability
(combined abiotic/SDM and biotic/host) spatial layer.}

\item{arrival_likelihood_layer}{A \code{raster::RasterLayer} or
\code{terra::SpatRaster} object representing the pest arrival likelihood
spatial layer (a combination of one or more pathway layers).}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the pest establishment
likelihood spatial layer (combined pest suitability and arrival
likelihood).
}
\description{
Builds a pest establishment likelihood spatial layer by combining
(multiplying) pest suitability and arrival likelihood layers.
}
\note{
Generalized modified version of corresponding layer-combining
functions in \code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
