% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_likelihood.R
\name{pathway_likelihood}
\alias{pathway_likelihood}
\alias{pathway_likelihood.Raster}
\alias{pathway_likelihood.SpatRaster}
\title{Estimate pathway pest arrival likelihood}
\usage{
pathway_likelihood(
  pathway_layers,
  leakage_rate_ci,
  establishment_rate_ci,
  confidence = 0.95,
  filename = "",
  ...
)

\method{pathway_likelihood}{Raster}(
  pathway_layers,
  leakage_rate_ci,
  establishment_rate_ci,
  confidence = 0.95,
  filename = "",
  ...
)

\method{pathway_likelihood}{SpatRaster}(
  pathway_layers,
  leakage_rate_ci,
  establishment_rate_ci,
  confidence = 0.95,
  filename = "",
  ...
)
}
\arguments{
\item{pathway_layers}{A \code{raster::Raster*} or \code{terra::SpatRaster}
object containing one or more layers for estimating the spatial
distribution of arrival likelihood.}

\item{leakage_rate_ci}{Numeric vector of 2 values, giving the lower and
upper bounds (CI) for leakage rate (the number of pest leakage events in
a random year).}

\item{establishment_rate_ci}{Numeric vector of 2 values, giving the lower
and upper bounds (CI) for establishment rate (the rate of survival, to
establishment, for leakage events).}

\item{confidence}{Confidence interval (CI). Default = 0.95.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{raster::RasterLayer} or \code{terra::SpatRaster} object
(as per \code{pathway_layers}) containing estimated pathway pest arrival
likelihoods.
}
\description{
Estimates spatially-explicit pest arrival likelihoods for a given pathway
from overall leakage and establishment rate estimates and their likely
spatial distribution.
}
\note{
Generalized modified version of
\code{\href{edmaps}{https://github.com/jscamac/edmaps}::arrivals_by_*}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
