% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pest_suitability.R
\name{pest_suitability}
\alias{pest_suitability}
\title{Pest suitability risk mapping layer}
\usage{
pest_suitability(abiotic_layer, biotic_layer, filename = "", ...)
}
\arguments{
\item{abiotic_layer}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the abiotic (SDM) spatial layer.}

\item{biotic_layer}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the biotic (host) spatial layer.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the combined pest
suitability layer.
}
\description{
Builds a pest suitability spatial layer by combining (multiplying) abiotic
(SDM) and biotic (host) suitability layers.
}
\note{
Generalized modified version of corresponding layer-combining
functions in \code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
