% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_layer.R
\name{transform_layer}
\alias{transform_layer}
\alias{transform_layer.Raster}
\alias{transform_layer.SpatRaster}
\title{Transform spatial layer}
\usage{
transform_layer(
  x,
  type = c("linear", "exponential", "logarithmic", "lower", "upper"),
  a = NULL,
  b = NULL,
  filename = "",
  ...
)

\method{transform_layer}{Raster}(
  x,
  type = c("linear", "exponential", "logarithmic", "lower", "upper"),
  a = NULL,
  b = NULL,
  filename = "",
  ...
)

\method{transform_layer}{SpatRaster}(
  x,
  type = c("linear", "exponential", "logarithmic", "lower", "upper"),
  a = NULL,
  b = NULL,
  filename = "",
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer to be transformed.}

\item{type}{One of \code{"linear"} (\code{a*x + b}), \code{"exponential"}
(\code{a^x*x^b}), \code{"logarithmic"} (\code{log(x, base = b)}),
\code{"lower"} (\code{x[x < a] = b}), or \code{"upper"}
(\code{x[x > a] = b}).}

\item{a}{Numeric parameter for transformation as specified by \code{type}.
Default (\code{NULL}) implies: \code{1} for \code{"linear"},
\code{"exponential"}, and \code{"upper"}; and \code{0} for \code{"lower"}.}

\item{b}{Numeric parameter for transformation as specified by \code{type}.
Default (\code{NULL}) implies: \code{0} for \code{"linear"},
\code{"exponential"}, and \code{"lower"}; \code{1} for \code{"upper"}; and
\code{exp(1)} (natural) for \code{"logarithmic"}.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the transformed layer.
}
\description{
Transforms a spatial layer via a variety of operations, including the
application of: linear, exponential, or logarithmic expressions; or lower
or upper thresholds; to layer values.
}
