% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_layers.R
\name{combine_layers}
\alias{combine_layers}
\alias{combine_layers.Raster}
\alias{combine_layers.SpatRaster}
\title{Combine spatial layers}
\usage{
combine_layers(
  x,
  use_fun = c("prod", "sum", "union"),
  na.rm = FALSE,
  binarize = FALSE,
  filename = "",
  ...
)

\method{combine_layers}{Raster}(
  x,
  use_fun = c("prod", "sum", "union"),
  na.rm = FALSE,
  binarize = FALSE,
  filename = "",
  ...
)

\method{combine_layers}{SpatRaster}(
  x,
  use_fun = c("prod", "sum", "union"),
  na.rm = FALSE,
  binarize = FALSE,
  filename = "",
  ...
)
}
\arguments{
\item{x}{A multi-layer \code{raster::Raster*} or \code{terra::SpatRaster}
object representing the spatial layers to combine.}

\item{use_fun}{One of \code{"prod"}, \code{"sum"}, or \code{"union"}.}

\item{na.rm}{Logical indicating whether or not to ignore \code{"NA"} values
when applying the combination function (\code{"use_fun"}).}

\item{binarize}{Logical indicating if the combined cells should be
binarized, i.e. set to 1 for values > 0. Default = \code{FALSE}.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{raster::RasterLayer} or \code{terra::SpatRaster} object
(as per \code{pathway_layers}) containing the combined layer.
}
\description{
Combines multiple spatial layers via cell-wise multiplication, addition,
or union (via complements), and optionally binarizes the output.
}
\note{
Generalized modified version of corresponding layer-combining
functions in \code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
