% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conform_layer.R
\name{conform_layer}
\alias{conform_layer}
\alias{conform_layer.Raster}
\alias{conform_layer.SpatRaster}
\title{Conform spatial layer}
\usage{
conform_layer(x, y, normalize = FALSE, binarize = FALSE, filename = "", ...)

\method{conform_layer}{Raster}(x, y, normalize = FALSE, binarize = FALSE, filename = "", ...)

\method{conform_layer}{SpatRaster}(x, y, normalize = FALSE, binarize = FALSE, filename = "", ...)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer to be conformed.}

\item{y}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial layer to conform to.}

\item{normalize}{Logical indicating if the conformed cells should be
normalized, i.e. set to a value 0-1 based on cell-wise minimum and
maximum values, i.e. \code{(value - min)/(max - min)}. Default =
\code{FALSE}.}

\item{binarize}{Logical indicating if the combined cells should be
binarized, i.e. set to 1 for values > 0. Default = \code{FALSE}.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing the conformed layer.
}
\description{
Conforms the spatial configuration (CRS, extent, and resolution) of a
spatial layer to that of another (template) layer.
}
\note{
Informed by various functions in
\code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
