% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weight_layer.R
\name{distance_weight_layer}
\alias{distance_weight_layer}
\alias{distance_weight_layer.Raster}
\alias{distance_weight_layer.SpatRaster}
\title{Calculate a distance weight layer}
\usage{
distance_weight_layer(
  x,
  y,
  beta = log(0.5)/200,
  weights = NULL,
  max_distance = NULL,
  filename = "",
  ...
)

\method{distance_weight_layer}{Raster}(
  x,
  y,
  beta = log(0.5)/200,
  weights = NULL,
  max_distance = NULL,
  filename = "",
  ...
)

\method{distance_weight_layer}{SpatRaster}(
  x,
  y,
  beta = log(0.5)/200,
  weights = NULL,
  max_distance = NULL,
  filename = "",
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial configuration (template) for calculating
the distance weight layer from points \code{y}, or a for a pre-calculated
distance layer (assumed when \code{y} is missing).}

\item{y}{Point (feature) data as a \code{data.frame} (or \code{matrix})
with WGS84 \emph{lon} and \emph{lat} columns, and optionally a weights
column having character name specified in \emph{weights}.}

\item{beta}{Numeric parameter passed to the exponential function:
\code{exp(distance/(1000/beta))}, which transforms cell-to-nearest-point
or pre-calculated distances (m) to distance-weighted cell values. To
generate a distribution that ensures a proportion \emph{p} of a pathway
likelihood is distributed within distance of each point (feature) or
pre-calculated distance \emph{d}, specify \code{beta=log(p)/d} (e.g. to
have 50\% of the likelihood distributed within 200 km use default:
\code{log(0.5)/200}).}

\item{weights}{Optional numeric vector of weights for each point (feature),
or character name of column or attribute in \code{y}, that contains weight
values for each point, such that the distance-weighted cell values will be
calculated via weighted sums. Thus the exponential function (above)
becomes: the sum of \code{exp(distance/(1000/beta))*weight} for each cell.
Default is none (\code{NULL}), in which case the distance to the nearest
point (feature) is used in the exponential function.}

\item{max_distance}{Optional maximum distance (m) used to calculate the
distance weight layer. Default is \code{NULL}, which considers all
distances.}

\item{filename}{Optional file writing path (character).}

\item{...}{Additional parameters (passed to \code{writeRaster}).}
}
\value{
A \code{terra::SpatRaster} object containing distance-weighted cell
values.
}
\description{
Calculates a distance-weighted spatial layer based on the proximity of each
cell to a series of points (features), or via a pre-calculated distance
layer. Each cell is weighted via a negative exponential function for
calculating values that decay with distance from each point. Optional
weights may be assigned to the points (features), in which case the
proximity values are calculated at each cell via weighted sums, otherwise
they are calculated using the nearest point (feature) to each cell.
}
\note{
Generalized modified version of corresponding distance-weight
functions in \code{\href{edmaps}{https://github.com/jscamac/edmaps}}.
}
\references{
Camac, J. & Baumgartner, J. (2021). \emph{edmaps} (early
detection maps) : An R package for creating Australian maps of
establishment likelihood for terrestrial plant pests.
\url{https://github.com/jscamac/edmaps}.
}
